+++ prisma/migrations/20260916021747_add_purchase_revocation/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "ConsumedPurchase" ADD COLUMN "revokedAt" DATETIME;
+
+-- AlterTable
+ALTER TABLE "User" ADD COLUMN "disabledAt" DATETIME;